requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
passlib==1.7.4
bcrypt==3.2.2
python-jose==3.3.0
mysql-connector-python
pytest==7.4.3
